oauth/requirements.txt
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudevents==1.11.0
cryptography==44.0.1
deprecation==2.1.0
Flask==3.0.2
functions-framework==3.8.2
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
googleapis-common-protos==1.67.0
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jwt==1.3.1
MarkupSafe==3.0.2
packaging==24.2
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
urllib3==2.3.0
watchdog==6.0.0
Werkzeug==3.1.3
